backend/requirements.txt
Flask
flask_cors
pymilvus
python-dotenv
requests
watchdog
gunicorn
click
python-docx
beautifulsoup4
fastapi
uvicorn
python-multipart
httpx
pydantic
mammoth
docxtpl
Pillow
pandas
openpyxl
google-genai